refactor: migrate styles to src dir
